refactor: 각 도메인 컴포넌트로 분리
